Updated flux calculation
The previous version contained an error and the reported flux densities
were in Jansky/pixel, not Jansky/beam as stated. This update fixes the
issue, and flux densities are now in Jansky/beam.

Furthermore, the radius of the emitting region (or outflow lobe) now
corresponds to the devonvolved size of the emitting region as observed
with the JCMT.